DATAREDIS-257 - Shutdown the client when DefaultLettucePool is destroyed.
DefaultLettucePool manages a RedisClient instance that's created
when afterPropertiesSet is called. The client was not being shutdown
as part of the pool's destroy() processing leading to a resource leak.
On OS X this would manifest itself as a growing number of kqueue file
descriptors being consumed eventually leading to a failure when
something tries to get a file descriptor and is unable to do so.

This commit updates destroy() to call the client's shutdown() method.

Original pull request: #26